Improve efficiency of skipping arrays we're not interested in in input data
This effectively puts the array skip code back to how it was in 1.5 for arrays
we're not interested in.  When code was added to 2.0 to make it possible to see
the values of array fields (primarily for lat_long), skipping of uninteresting
array fields became considerably less efficient and potentially wrong.


Former-commit-id: b451adf74826f08257eff549ee38ae49c1dc5972

Original commit: elastic/x-pack@9db21574e92ae8de976f33f04f8890bc66c8b54d