Refactor serialisation
Currently the serialisation of objects that are
persisted in the storage is done in ElasticsearchPersister.

This results to the following disadvantages:

  - The ElasticsearchPersister is too big.
  - Adding a new field in a data type is error prone
  as one has to remember to add it in its serialise method
  in ElasticsearchPersister.
  - Implementing a new storage connector will require rewriting
  all this logic when the only different bit is the actual
  implementation of the serialisation and not the mapping.
  - Unit testing all the serialisation mapping logic is impossible.

This commit refactors all that. Each object that gets serialised
implements a StorageSerialisable interface. The logic is moved
into each object's serialise method. An implementation for
Elasticsearch is added. Also, I add an implementation that
serialises to JSON to facilitate unit testing.

Unit tests are added for all objects.


Former-commit-id: c437cd7710185e8f48452606d1bc6124add1ef77

Original commit: elastic/x-pack@acdfe356cf04ab7c19fc4685495d4496d02627b4