security: add metadata to roles
This commit adds the ability to define metadata for roles. This metadata is currently
only used for the API and to indicate that a role is reserved. We can continue passing
on the metadata as needed, when necessary.

Closes elastic/x-pack#2036

Original commit: elastic/x-pack@2aca98d997a88dbe52fa5d27eec0a4c4655455c1