SS16-116 Add the average bucket processing time to the job
Average is the exponential moving average it’s stored in a different document type so as not to cause contention with the job doc when it is updated. The API merges the value into the job doc on request.

Former-commit-id: c60b6e83ad5da2fe2d531e66a4e35a2c6c3f9290

Original commit: elastic/x-pack@b81b5e156d6d3b9b002b2146be3a10a25d480de8