security: do not use hidden filenames when generating certs
This commit changes how we get the file and directory name for certificates in the tool. The
tool now prompts the user for the filename. If the provided instance name will result in a
valid filename, this is provided as a default. Otherwise the user must provide a valid
filename.

Closes elastic/x-pack#2854

Original commit: elastic/x-pack@eb03418c52b8ef47a5f23281eab2681e2968fd02