security: allow enabled and username fields in put user request body
The enabled and username fields are both now allowed in the request body for the put user
request. This makes it easier to perform a get and update a user without needing to edit more
of the request body than necessary.

Closes elastic/x-pack#3391

Original commit: elastic/x-pack@bb62cf8afebf800dc67aef0267d6234305e5269b