Explicitly shut down watcher after tests
In our tests you have to explicitly shut down watcher rather than shut
down the node it is running on because of thread leak detection. Just
shutting down the node that it is running on will cause it to start up
on another node if there is another one running and then not properly
shut down. This is probably something that should be fixed in watcher
somehow but for now lets just be more careful with the tests.

Closes elastic/x-pack#2365
Closes elastic/x-pack#2588

Original commit: elastic/x-pack@9803b1d43afadcfdaf4df7a0a00f017c5dabf1da