test: stop IndexAuditTrail before ESIntegTestCase#after runs
As part of the review of elastic/x-pack#3287, the stopping of the IndexAuditTrail was moved to the tearDown
method. This works sometimes but other times it fails because tearDown is run after
ESIntegTestCase#after, so the IndexAuditTrail is still running during the after checks which will
cause the test to fail since the shard lock cannot be obtained.

Closes elastic/x-pack#3520

Original commit: elastic/x-pack@fde7e344b854eb83e46944ff1e278e642772c37c