Watcher: Ignore watch version when deleting a watch (elastic/x-pack#3630)
When deleting a watch the version was used as part of
the delete request. However a watch deletion means the
user really wants to get rid of it and not accidentally
run into a version exception because the watch was running in
between.

Original commit: elastic/x-pack@9d1240ae1467e1d6f59de4e84b0cc796fb7a2676