security: native users and roles can be used on tribe nodes
This change allows native users and roles to be used on tribe nodes. The tribe node will actually
use the security index of one of the tribes, which must be specified with the `tribe.on_conflict`
setting. User and role modifications are not permitted when running on a tribe node.

Closes elastic/x-pack#3451

Original commit: elastic/x-pack@37cfd764f21d2a6bbf8bcc1049d3aad0509caa42