Add Vagrant build environment (elastic/x-pack#38)
This adds a vagrant configuration under /vagrant. This allows the user to provision a virtual machine with the entire build environment needed for compiling the C++ portion of the code (as well as the Java side, but that can be done easily outside of vagrant).

Original commit: elastic/x-pack@2e6a98307c662d8a4cffde133b4c404e32b26cb6