Advance scheduled job in proper real-time fashion
This changes the behaviour of a real-time scheduled job
to keep up with real-time regardless of whether data is seen
or not. The reason for that is because it ensures anomalies will
be created for the case that the absence of data is the anomaly.
It also makes the behaviour of the scheduler easier to explain
and predict. Finally, it makes the code simpler.

This commit also contains 2 more refactorings:
  - The extractor logic has been rewritten in terms of the time
    advancement into what seems to be much simpler to argue for.
  - Flushing was changed to not throw interrupted exception in order
    to avoid problems when the scheduler is manually stopped while it
    is performing a flush. This is particularly possible with the new
    real-time behaviour as while advancing time for historical data
    to now there can be a big gap that forces a long flush.

Original commit: elastic/x-pack@00998cdf5ba13276842e49d8f2e98ea0cb825467