Remove Notification Accounts abstraction (elastic/x-pack#3811)
This change is a first step towards a real abstraction on top of all the
notification services. There are a bunch of followup changes coming for this
that will remove most of the classes in here but this is a first small step
to actually have a notification service interface.

Original commit: elastic/x-pack@ab921dfa4f6f67dc27011d007169ac88ea17b8b3